refactor: 리팩터링 요구 사항에 따라 도메인 로직은 domain 폴더, UI 관련 로직은 view 폴더에서 관리하기 위해 폴더명 수정
